Add files via upload
In this game your program generates a random number between 1 and 100. You need to find that number. The first thing that you have to do is choosing the difficulty of the game. If you choose easy version than you have 7 steps to find the generated number, in medium version you have 5 and in hard version you have 3 steps to find the number. If you exceed the limit of steps than you win otherwise you need to play the game all over again